Name jsons as the accession
This commit introduce a feature where the generated json files
are getting named like the migrated accessions.
It also changes the default behaviour of the processed key_list.
Now there is a hardcoded key_list which will function as a default
parameter. If one ones to specify another keylist this can still be
done through the parameter - although thats not to handy yet. Maybe
it should be considered if we want to pass a file with the keylist
rather than the plain keylist but this has to be discussed as a
usecase.