Fix issue in relabeling json
This commit fixes an issue which results in empty dicts caused
by the relabel_json function which was not working properly.
The lowering of the keys and values resulted in a problem where
the file_as_dict was a generator instead of a dict so that the
if clause was never executed and thus empty dicts returned.